Revert "Прошли тесты постман для всей сборки"
This reverts commit fa18b072adf86967778fe9d8a2c281c2be020607.

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -3,42 +3,37 @@ package ru.yandex.practicum.filmorate.service;
 import java.time.LocalDate;
 import java.util.Collection;
 import java.util.List;
-import java.util.Set;
+import java.util.Optional;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import ru.yandex.practicum.filmorate.dal.*;
 import ru.yandex.practicum.filmorate.exception.BadRequestException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.model.*;
+import ru.yandex.practicum.filmorate.model.Event;
+import ru.yandex.practicum.filmorate.model.EventType;
+import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.Operation;
 
 @Slf4j
 @Service
-@Transactional
 @RequiredArgsConstructor
 public class FilmService {
 
     private final FilmRepository filmRepository;
-    private final UserService userService;
+    private final UserRepository userRepository;
     private final DirectorRepository directorRepository;
     private final MpaService mpaService;
     private final GenreRepository genreRepository;
-    private final EventService eventService;
+    private final EventRepository eventRepository;
 
     public void removeFilm(int id) {
         filmRepository.removeFilmById(id);
     }
 
     public Collection<Film> findAll() {
-        Collection<Film> films = filmRepository.findAll();
-
-        for (Film film : films) {
-            enrichFilmWithData(film);
-        }
-
-        return films;
+        return filmRepository.findAll();
     }
 
     public Film create(Film film) {
@@ -64,48 +59,69 @@ public class FilmService {
 
         film = filmRepository.save(film);
 
-        updateGenres(film, false);
-        updateDirectors(film, false);
+        if (film.getGenres() != null && !film.getGenres().isEmpty()) {
+            genreRepository.addGenre(film);
+        }
 
         return film;
     }
 
     public Film update(Film newFilm) {
-        if (newFilm.getId() == null || !filmRepository.existsById(newFilm.getId())) {
-            throw new NotFoundException("Фильм с id " + newFilm.getId() + " не найден");
+        if (newFilm.getId() == null || filmRepository.findById(newFilm.getId()).isEmpty()) {
+            throw new NotFoundException("Фильм с id = " + newFilm.getId() + " не найден");
         }
 
-        filmRepository.update(newFilm);
+        newFilm = filmRepository.update(newFilm);
 
-        updateGenres(newFilm, true);
-        updateDirectors(newFilm, true);
+        if (newFilm.getGenres() != null && !newFilm.getGenres().isEmpty()) {
+            genreRepository.updateGenre(newFilm);
+        }
 
-        return findById(newFilm.getId());
+        return newFilm;
     }
 
-    public Film findById(Integer id) {
-        Film film = filmRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Фильм с id " + id + " не найден"));
-
-        enrichFilmWithData(film);
-
-        return film;
+    public Optional<Film> findById(Integer id) {
+        return filmRepository.findById(id);
     }
 
     public void likeTheMovie(Integer filmId, Integer userId) {
-        throwIfNotExists(filmId);
-        userService.throwIfNotExists(userId);
+        if (filmRepository.findById(filmId).isEmpty()) {
+            throw new NotFoundException("Фильм не найден");
+        }
+
+        if (userRepository.findById(userId).isEmpty()) {
+            throw new NotFoundException("Пользователь не найден");
+        }
+
         filmRepository.likeFilm(filmId, userId);
-        eventService.saveEvent(userId, filmId, EventType.LIKE, Operation.ADD);
+        Event event = new Event(); // добавление в ленту
+        event.setEventType(EventType.LIKE.toString());
+        event.setUserId(userId); // юзер добавил лайк фильму
+        event.setEntityId(filmId);
+        event.setOperation(Operation.ADD.toString());
+        eventRepository.save(event); // добавление в ленту
     }
 
     public void removeLikeTheMovie(Integer filmId, Integer userId) {
-        throwIfNotExists(filmId);
-        userService.throwIfNotExists(userId);
+        if (filmRepository.findById(filmId).isEmpty()) {
+            throw new NotFoundException("Фильм не найден");
+        }
+
+        if (userRepository.findById(userId).isEmpty()) {
+            throw new NotFoundException("Пользователь не найден");
+        }
 
         filmRepository.removeLike(filmId, userId);
+        Event event = new Event(); // добавление в ленту
+        event.setEventType(EventType.LIKE.toString());
+        event.setUserId(userId); // юзер удалил лайк фильму
+        event.setEntityId(filmId);
+        event.setOperation(Operation.REMOVE.toString());
+        eventRepository.save(event); // добавление в ленту
+    }
 
-        eventService.saveEvent(userId, filmId, EventType.LIKE, Operation.REMOVE);
+    public List<Film> getFilmWithTheMostLikes(Integer count) {
+        return filmRepository.findMostLikedFilms(count);
     }
 
     public List<Film> findAllFilmByDirector(Integer directorId, String sortBy) {
@@ -113,31 +129,17 @@ public class FilmService {
             throw new NotFoundException("Режиссёр не найден");
         }
 
-        List<Film> films = null;
-
         if ("likes".equals(sortBy)) {
-            films = filmRepository.findFilmsByDirectorSortedByLikes(directorId);
+            return filmRepository.findFilmsByDirectorSortedByLikes(directorId);
         } else if ("year".equals(sortBy)) {
-            films = filmRepository.findFilmsByDirectorSortedByYear(directorId);
+            return filmRepository.findFilmsByDirectorSortedByYear(directorId);
         } else {
             throw new IllegalArgumentException("Unknown sortBy value: " + sortBy);
         }
-
-        for (Film film : films) {
-            enrichFilmWithData(film);
-        }
-
-        return films;
     }
 
     public List<Film> getCommonSortedFilms(Integer userId, Integer friendId) {
-        List<Film> films = filmRepository.getCommonSortedFilms(userId, friendId);
-
-        for (Film film : films) {
-            enrichFilmWithData(film);
-        }
-
-        return films;
+        return filmRepository.getCommonSortedFilms(userId, friendId);
     }
 
     public List<Film> getPopular(Integer count, Long genreId, Integer year) {
@@ -150,79 +152,18 @@ public class FilmService {
                     .orElseThrow(() -> new NotFoundException("Жанр с id = " + genreId + " не найден"));
         }
 
-        List<Film> films = filmRepository.findMostPopularsByGenreAndYear(count, genreId, year);
-
-        for (Film film : films) {
-            enrichFilmWithData(film);
-        }
-
-        return films;
+        return filmRepository.findMostPopularsByGenreAndYear(count, genreId, year);
     }
 
     public List<Film> findByTitleOrDirector(String query, String by) {
-
         String[] result = by.split(",");
-        List<Film> films = null;
 
         if (result.length == 2) {
-            films = filmRepository.findByTitleAndDirector(query);
+            return filmRepository.findByTitleAndDirector(query);
         } else if (result[0].equals("title")) {
-            films = filmRepository.findByTitle(query);
+            return filmRepository.findByTitle(query);
         } else if (result[0].equals("director")) {
-            films = filmRepository.findByDirector(query);
+            return filmRepository.findByDirector(query);
         } else throw new BadRequestException("Неверно указаны параметры строки запроса");
-
-        for (Film film : films) {
-            enrichFilmWithData(film);
-        }
-
-        return films;
-    }
-
-    public void throwIfNotExists(Integer userId) {
-        if (!filmRepository.existsById(userId)) {
-            throw new NotFoundException("Фильм с id = " + userId + " не найден");
-        }
-    }
-
-    public Collection<Film> getUniqueMovies(Integer sourceUserId, Integer targetUserId) {
-        Collection<Film> films = filmRepository.getUniqueMovies(sourceUserId, targetUserId);
-
-        for (Film film : films) {
-            enrichFilmWithData(film);
-        }
-
-        return films;
-    }
-
-    private void updateGenres(Film film, boolean isUpdate) {
-        if (film.getGenres() != null) {
-            Set<Genre> genres = Set.copyOf(film.getGenres());
-            if (isUpdate) {
-                genreRepository.deleteGenresFromFilm(film.getId());
-            }
-            genreRepository.updateFilmGenres(film.getId(), genres);
-        }
-    }
-
-    private void updateDirectors(Film film, boolean isUpdate) {
-        if (film.getDirectors() != null) {
-            Set<Director> directors = Set.copyOf(film.getDirectors());
-            if (isUpdate) {
-                directorRepository.deleteFilmDirectors(film.getId());
-            }
-            directorRepository.saveFilmDirectors(film.getId(), directors);
-        }
-    }
-
-    // меняет объект по ссылке
-    private void enrichFilmWithData(Film film) {
-        film.setGenres(genreRepository.getGenresByFilmId(film.getId()));
-
-        if (film.getMpa() != null) {
-            film.setMpa(mpaService.findById(film.getMpa().getId()));
-        }
-
-        film.setDirectors(directorRepository.getDirectorsByFilmId(film.getId()));
     }
 }
